requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg==3.1.18
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
stripe==9.3.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
